routing/tests: Use newly created graphSource
This commit shifts old tests to use the new pathfinding functions.
It also adds a benchmark test for a call to findPaths. The
graph used is the testnet graph from mid-October 2017, and an utility
method for parsing the JSON produced by lncli describegraph is also
included.

The bechmark test  was clocking in at approx 11 seconds with the
ChannelGraph as the graph passed to findPath. This is now reduced
to sub 50 ms with the previous commmit, which introduced an in-
memory graph.

Co-authored-by: Johan T. Halseth <johanth@gmail.com>