routing: Create graph interface and an in-memory graph
This commit creates a graph interface graphSource for use in pathfinding
algorithms.
It only provides basic features of iteration on graph nodes and edges.

A struct Edge is defined which stores all information relevant to
pathfinding.
Struct memChannelGraph which implements graphSource is
defined. It can be created using a ChannelGraph and stores everything
relevant for pathfinding in-memory, so use/access is much faster after
initial creation.

Finally, dbChannelGraph, a wrapper for channeldb.ChannelGraph is defined.